Switch to Vite build system (#29)
`react-scripts` and `create-react-app` are deprecated. The current suggestion is to switch to `vite`. This PR switches to `vite` for building and development.

Since we are not serving a SPA and all of our content is compiled outside of pretext-react, some cleverness is required to work with vite. A plugin for serving sources in dev mode was developed and allows hot-reloading.